Only check coverage of unit tests
