+++ test/e2e/TestConfig.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { mockConfig } from './helpers';
+import { mockConfig } from '../helpers';
 
 if (process.env.RUN_ONLY_OPTIONAL_TESTS !== 'true') {
   describe('Config', () => {
@@ -7,5 +7,5 @@ if (process.env.RUN_ONLY_OPTIONAL_TESTS !== 'true') {
       const config = mockConfig();
       expect(config).to.exist;
     }).timeout(10000);
-  })
+  });
 }
