+++ test/unit/dependencies.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 
 // from https://docs.npmjs.com/files/package.json#dependencies
 const nonExactPrefixes = ['~', '^', '>', '>=', '<', '<='];
-const packageJSON = require('../package.json');
+const packageJSON = require('../../package.json');
 
 describe('package.json', () => {
   it('dependencies should not contain any non-exact versions', () => {
@@ -14,7 +14,9 @@ describe('package.json', () => {
     });
   });
   it('devDependencies should not contain any non-exact versions', () => {
-    const deps = Object.keys(packageJSON.devDependencies).map(key => packageJSON.devDependencies[key]);
+    const deps = Object.keys(packageJSON.devDependencies).map(
+      key => packageJSON.devDependencies[key]
+    );
     deps.forEach(depVersion => {
       nonExactPrefixes.forEach(badPrefix => {
         expect(depVersion.includes(badPrefix)).to.be.false;
